Fix FPS drop when joining voice chat
- Throttle AudioEngine GetTelemetry() to every 100ms when in voice; use cached value for redundancy and stats to avoid m_TelemetryMutex contention
- Use waitMs=0 when in voice channel (or target FPS set) so message loop is never throttled by 250ms/48ms
- Skip overlay GDI+ Repaint when member list unchanged; throttle overlay repaints to ~15 FPS max to avoid main-loop stall in voice

Made-with: Cursor

+++ src/screen/DXGICapture.cpp
@@ -189,8 +189,8 @@ bool CaptureWindowContentToBgra(HWND hwnd, int outW, int outH, uint8_t* outBgra)
     if (!ok) {
         RECT wr = {};
         ::GetWindowRect(hwnd, &wr);
-        const int srcW = (std::max)(1, wr.right - wr.left);
-        const int srcH = (std::max)(1, wr.bottom - wr.top);
+        const int srcW = (std::max)(1, static_cast<int>(wr.right - wr.left));
+        const int srcH = (std::max)(1, static_cast<int>(wr.bottom - wr.top));
         ::SetStretchBltMode(memDc, HALFTONE);
         ok = ::StretchBlt(memDc, 0, 0, outW, outH, windowDc, 0, 0, srcW, srcH, SRCCOPY);
     }
